Generate REST resources and models from an OpenAPI spec
This commit adds an example OpenAPI specification (the default Pets API)
to the project and sets up the Maven OpenAPI generator plugin to
generate API interfaces and models that are used for implementing the
API inside the Quarkus project.

The general configuration tweaks are documented inline. A few quirks of
this setup will be fixed in the following commits.