Only expose endpoints declared in the static OpenAPI specification
By default, Quarkus serves an OpenAPI specification that is the
combination of our static OpenAPI file and additional routes only
declared in the application source code. In an API-first world, we
usually don't want to expose anything (or document using SwaggerUI) that
has not been negotiated using the API-first principles. Therefore, this
commit disables adding discovered endpoints to the served OpenAPI
specification and SwaggerUI.