Validate spec compliance with prism proxy
Adds a Quarkus test resource launching stoplight prism as a docker
container, configured with the Quarkus app as the upstream and the
OpenAPI specification present. This proxy is used in tests instead of
the actual app so that each request and response towards the application
is validated for spec compliance.

The new checks have uncovered a configuration issue in the Jackson
serialization. Null values were serialized as null, even though this was
not valid due `nullable` being false for the affected fields.